[component] Service-Component & wildcards
PDE projects tend to have a Service-Component
header like "OSGI-INF/*.xml", which is valid
SCR header. However, this clashes with the bnd
generation since it will append the xml file paths
to the end of this header. SCR will then see 
components twice. Once matched by the wildcard
and once literal.

Unfortunately, many PDE projects are in progress
to moving the annotations based components so they 
have a hybrid situation. This makes it impossible 
for them to remove the Service-Component wildcard 
since then no longer the hardcoded files are found.

Rock and a hard place.

So this patch will match the calculated paths
in the old ServiceComponent class  and the more
modern DSAnnotations class against any wildcard
patterns. The static method that does this is located
in DSAnnotations.

Sucks but after trying to find alternatives this turned
out to be the best solution.






Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>