[#4858] UI to add bundles was confusing
If you searched for a bundle in the bndeditor/build tab
then it would not list any matches that were already
added to the build path. This could give the false
impression that the bundle was not in the list,
this confusion caused #4858

This patch will always list any matching bundles.
However, it will filter any bundles that are already
part of the selected bundles. The addButton is
also only enabled if there is at least 1 bundle
selected in the available list that is not in the
selected list.

Fixes #4858

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>